audio_service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.7
openai==1.86.0
python-dotenv==1.1.0
